# contracts/openapi.yaml
...
openapi: 3.1.0
info:
  title: SoulAuth
  version: "0.1.0"
  description: >-
    Actor-native identity and authentication infrastructure.
    Endpoint surface only; payload schemas are not yet frozen.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: "{APP_URL}"
    description: >-
      The public address, set by APP_URL. It is also the OIDC issuer — the two
      must match character for character.
    variables:
      APP_URL:
        default: http://localhost:8080
components:
...
  schemas:
    Error:
      type: object
      required:
      - error
      - message
      additionalProperties: true
      description: 'The single error envelope used across the whole API. `error` is a **stable machine-readable
        code** — branch on it. `message` is prose: its wording may improve at any time and must not be
        used for decisions. Individual errors carry documented additional fields.'
      properties:
        error:
          type: string
          description: Stable error code, snake_case.
          enum:
          - internal_error
          - invalid_credentials
          - email_not_verified
          - invalid_token
          - user_not_found
          - email_exists
          - username_exists
          - oauth_error
          - password_already_set
          - invalid_user_id
          - not_found
          - validation_error
          - permission_denied
          - insufficient_permissions
          - missing_permission
          - account_suspended
          - account_inactive
          - account_deleted
          - account_locked
          - rate_limit_exceeded
          - forbidden
          - bad_request
          - not_configured
          - unauthorized
          - service_unavailable
        message:
          type: string
          description: Human-readable explanation. Not part of the contract.
        required_permission:
          type: string
          description: Carried by `missing_permission` only — which permission is missing.
        locked_until_seconds:
          type: integer
          description: Carried by `account_locked` only — seconds until the lock clears.
...
    ActorResponse:
      properties:
        actor_kind:
          type: string
        created_at:
          type: integer
        id:
          type: string
        identity_source:
          type: string
        status:
          type: string
        subject_key:
          type: string
      required:
      - id
      - subject_key
      - actor_kind
      - identity_source
      - status
      - created_at
      type: object
...
    AuthenticateRequest:
      properties:
        actor_id:
          type: string
        algorithm:
          type: string
        nonce:
          type: string
        signature:
          type: string
      required:
      - actor_id
      - nonce
      - algorithm
      - signature
      type: object
    AuthenticateResponse:
      properties:
        actor_id:
          type: string
        credential_label:
          type: string
        expires_at:
          type: integer
        token:
          type: string
        token_type:
          type: object
      required:
      - token
      - token_type
      - actor_id
      - expires_at
      - credential_label
      type: object
...
    RegisterRequest:
      properties:
        label:
          type: string
        public_key:
          type: string
      required:
      - public_key
      - label
      type: object
    RegisterResponse:
      properties:
        actor:
          $ref: '#/components/schemas/ActorResponse'
        credential:
          $ref: '#/components/schemas/CredentialResponse'
      required:
      - actor
      - credential
      type: object
...
  securitySchemes:
    # 三种，不是一种。
    #
    # 把它们混成一个 `bearerAuth` 会产生具体的误导：读契约的人会以为
    # /api/oidc/userinfo 收的是登录拿到的那枚会话令牌 —— 它收的是 OIDC
    # 访问令牌，两者由不同端点签发、生命周期也不同。
    bearerAuth:
      type: http
      scheme: bearer
      description: >-
        Session token, from POST /api/auth/login (or POST
        /api/actors/authenticate for an AI actor). Used by every /api business
        endpoint.
    oidcAccessToken:
      type: http
      scheme: bearer
      description: >-
        OIDC access token, from POST /api/oidc/token. Used by
        /api/oidc/userinfo **only**. Not interchangeable with the session token
        above.
    browserSession:
      type: apiKey
      in: cookie
      name: soulauth_session
      description: >-
        Browser session cookie. /api/oidc/authorize uses it to decide whether
        this browser is logged in; its absence is a redirect to the login page,
        not an error.
...
paths:
...
  /api/actors:
    get:
      tags:
      - AI Actors
      operationId: actors_list_actors
      description: >-
        List registered AI actors. Returns identity records only; credentials are fetched per
        actor.
      description_zh: >-
        列出已注册的 AI 主体。只返回身份记录，凭证需按主体单独获取。
      security:
      - bearerAuth: []
      x-required-permissions:
      - soulauth:actors.read
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ActorResponse'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
      - AI Actors
      operationId: actors_register
      description: >-
        Register a non-human actor and attach its first Ed25519 verification key. No
        HumanAccount is created — an AI actor has no email, username or password.
      description_zh: >-
        注册一个非人主体并为它挂上第一枚 Ed25519 验证密钥。不创建 HumanAccount：AI 主体没有邮箱、用户名或口令。
      security:
      - bearerAuth: []
      x-required-permissions:
      - soulauth:actors.write
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
  /api/actors/:actor_id:
    get:
      tags:
      - AI Actors
      operationId: actors_get_actor
      description: >-
        One AI actor's identity record.
      description_zh: >-
        返回单个 AI 主体的身份记录。
      security:
      - bearerAuth: []
      x-required-permissions:
      - soulauth:actors.read
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActorResponse'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
      - name: actor_id
        in: path
        required: true
        schema:
          type: string
...
  /api/actors/authenticate:
    post:
      tags:
      - AI Actors
      operationId: actors_authenticate
      description: >-
        Step two: submit an Ed25519 signature over the challenge in exchange for a session
        token. The signed content is recomputed independently by the server, and the challenge
        is consumed **before** verification.
      description_zh: >-
        第二步：提交对挑战的 Ed25519 签名，换取会话令牌。被签名内容由服务端独立重算，且挑战在验签**之前**就被消费。
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticateResponse'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthenticateRequest'
  /api/actors/challenge:
    post:
      tags:
      - AI Actors
      operationId: actors_challenge
      description: >-
        Step one: obtain a single-use challenge. Public. An unknown actor and a suspended actor
        return the same `invalid_credentials`, so this cannot become a channel for enumerating
        subjects.
      description_zh: >-
        第一步：领取一次性挑战。公开端点。未知主体与已停用主体返回相同的 `invalid_credentials`，因此它无法成为枚举主体的信道。
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssuedChallenge'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChallengeRequest'
  /api/actors/me:
    get:
      tags:
      - AI Actors
      operationId: actors_me
      description: >-
        Agent introspection. A human session token is explicitly refused with 403.
      description_zh: >-
        AI 主体自省。人类会话令牌会被以 403 明确拒绝。
      security:
      - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActorResponse'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/auth/admin/login:
    post:
      tags:
      - Authentication
      operationId: auth_admin_login
      description: >-
        Login for the administration console. Identical to the normal login, plus a check that
        the account has console access.
      description_zh: >-
        管理控制台登录。流程与普通登录相同，额外校验该账号是否具备控制台访问权。
      security: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
...
  /api/auth/initialize-password:
    post:
      tags:
      - Authentication
      operationId: auth_initialize_password
      description: >-
        Set a password on an account that has none — typically one created through federated
        login.
      description_zh: >-
        为尚无口令的账号设置口令，通常用于经联邦登录创建的账号。
      security:
      - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InitializePasswordRequest'
  /api/auth/login:
    post:
      tags:
      - Authentication
      operationId: auth_login
      description: >-
        Exchange email and password for a session token. Unknown account and wrong password
        return the same error with the same timing. With MFA enabled, returns a short-lived
        challenge token instead of a session.
      description_zh: >-
        用邮箱与口令换取会话令牌。账号不存在与口令错误返回相同的错误、耗时也相同。启用 MFA 时返回的是短期挑战令牌而非会话。
      security: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
...
  /api/auth/me:
    get:
      tags:
      - Authentication
      operationId: auth_get_current_user
      description: >-
        The account behind the presented session token.
      description_zh: >-
        返回所出示会话令牌背后的账户。
      security:
      - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/auth/mfa/login-verify:
    post:
      tags:
      - Authentication
      operationId: auth_mfa_login_verify
      description: >-
        Second step of MFA login: exchange the challenge token plus a TOTP or backup code for a
        session. Public, because the challenge token is the credential here.
      description_zh: >-
        MFA 登录的第二步：用挑战令牌加上 TOTP 验证码或备用码换取会话。公开端点，因为此处的凭证正是那枚挑战令牌。
      security: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MfaLoginVerifyRequest'
...
  /api/auth/register:
    post:
      tags:
      - Authentication
      operationId: auth_register
      description: >-
        Create a human account. Enforces the password policy and, when email verification is
        enabled, sends a verification link.
      description_zh: >-
        创建人类账户。执行口令策略；启用邮箱验证时会发送验证链接。
      security: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
  /api/auth/request-password-reset:
    post:
      tags:
      - Authentication
      operationId: auth_request_password_reset
      description: >-
        Start password recovery. Always reports success — reporting failure only for registered
        addresses would itself be an enumeration oracle. Any previous unused token for that
        address is invalidated.
      description_zh: >-
        发起口令找回。永远返回成功：只对已注册地址报错，本身就成了账号枚举的信道。该地址此前未使用的令牌会一并作废。
      security: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestPasswordResetRequest'
  /api/auth/resend-verification:
    post:
      tags:
      - Authentication
      operationId: auth_resend_verification
      description: >-
        Issue a new verification token and resend the link. Already-verified accounts are
        silently ignored, so this cannot be used as a mailer for arbitrary addresses.
      description_zh: >-
        签发新的验证令牌并重发链接。已验证的账号会被静默忽略，因此它无法被当作对任意地址发信的工具。
      security: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestPasswordResetRequest'
  /api/auth/reset-password:
    post:
      tags:
      - Authentication
      operationId: auth_reset_password
      description: >-
        Complete password recovery. The token is claimed atomically before the password is
        written, so two concurrent requests cannot both succeed. All the account's sessions are
        revoked.
      description_zh: >-
        完成口令找回。令牌在写入新口令之前被原子抢占，因此两个并发请求不可能同时成功。该账号的全部会话都会被吊销。
      security: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
...
  /api/auth/verify-email/:token:
    get:
      tags:
      - Authentication
      operationId: auth_verify_email
      description: >-
        Consume an email verification token. Public by necessity — requiring a token to verify
        an address would be circular. The token is claimed atomically, so it cannot be replayed.
      description_zh: >-
        消费一枚邮箱验证令牌。必须公开：要求先持有令牌才能验证地址是循环依赖。令牌被原子抢占，因此无法重放。
      security: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
      - name: token
        in: path
        required: true
        schema:
          type: string
...
  /api/users:
    get:
      tags:
      - Actors & Profiles
      operationId: user_management_list_users
      description: >-
        List accounts. Requires `soulauth:users.read`.
      description_zh: >-
        列出账户，需要 `soulauth:users.read` 权限。
      security:
      - bearerAuth: []
      x-required-permissions:
      - soulauth:users.read
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
      - name: page
        in: query
        required: false
        schema:
          type: integer
      - name: search
        in: query
        required: false
        schema:
          type: string
      - name: sort_by
        in: query
        required: false
        schema:
          type: string
      - name: sort_order
        in: query
        required: false
        schema:
          type: string
      - name: status
        in: query
        required: false
        schema:
          type: object
  /api/users/:user_id:
    get:
      tags:
      - Actors & Profiles
      operationId: user_management_get_user_by_id
      description: >-
        One account by id.
      description_zh: >-
        按 id 返回单个账户。
      security:
      - bearerAuth: []
      x-required-permissions:
      - soulauth:users.read
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
...